# specs/017-eliza-persona-editor/contracts/api.yaml
openapi: 3.0.3
info:
  title: AI Persona Editor API
  description: API contracts for full Eliza SDK persona editing
  version: 1.0.0

paths:
  /api/eliza/characters/{tokenId}:
    get:
      summary: Get AI character by WAGDIE token ID
      operationId: getCharacter
      parameters:
        - name: tokenId
          in: path
          required: true
          schema:
            type: string
          description: WAGDIE character token ID
      responses:
        '200':
          description: AI character found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AICharacter'
        '404':
          description: AI character not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      summary: Create or update AI character
      operationId: upsertCharacter
      security:
        - bearerAuth: []
      parameters:
        - name: tokenId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAICharacterInput'
      responses:
        '200':
          description: Character updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AICharacter'
        '201':
          description: Character created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AICharacter'
        '401':
          description: Unauthorized
        '403':
          description: Not character owner
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'

  /api/eliza/characters/{tokenId}/knowledge:
    get:
      summary: List knowledge documents for character
      operationId: listKnowledge
      security:
        - bearerAuth: []
      parameters:
        - name: tokenId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Knowledge documents list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/KnowledgeDocument'
        '401':
          description: Unauthorized
        '403':
          description: Not character owner

    post:
      summary: Upload knowledge document
      operationId: uploadKnowledge
      security:
        - bearerAuth: []
      parameters:
        - name: tokenId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: Text file (.txt or .md), max 50KB
      responses:
        '201':
          description: Document uploaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KnowledgeDocument'
        '400':
          description: Invalid file type or size
        '401':
          description: Unauthorized
        '403':
          description: Not character owner
        '409':
          description: Maximum documents reached (5)

  /api/eliza/characters/{tokenId}/knowledge/{documentId}:
    delete:
      summary: Delete knowledge document
      operationId: deleteKnowledge
      security:
        - bearerAuth: []
      parameters:
        - name: tokenId
          in: path
          required: true
          schema:
            type: string
        - name: documentId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Document deleted
        '401':
          description: Unauthorized
        '403':
          description: Not character owner
        '404':
          description: Document not found

  /api/eliza/characters/{tokenId}/export:
    get:
      summary: Export character as Eliza JSON
      operationId: exportCharacter
      security:
        - bearerAuth: []
      parameters:
        - name: tokenId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Eliza character JSON
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ElizaCharacterExport'
        '401':
          description: Unauthorized
        '403':
          description: Not character owner
        '404':
          description: Character not found

  /api/eliza/characters/{tokenId}/import:
    post:
      summary: Import Eliza character JSON
      operationId: importCharacter
      security:
        - bearerAuth: []
      parameters:
        - name: tokenId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ElizaCharacterExport'
      responses:
        '200':
          description: Character imported
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  imported:
                    type: array
                    items:
                      type: string
                    description: Fields imported into the character record
                  skipped:
                    type: array
                    items:
                      type: string
                    description: Fields skipped because they are backend-owned, unsupported, or route-managed
                  warnings:
                    type: array
                    items:
                      type: string
                    description: Fields that were ignored, skipped, or modified
        '401':
          description: Unauthorized
        '403':
          description: Not character owner
        '422':
          description: Invalid Eliza character format

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    AICharacter:
      type: object
      required:
        - id
        - externalId
        - name
        - bio
        - lore
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        externalId:
          type: string
          description: WAGDIE token ID
        name:
          type: string
          maxLength: 100
        username:
          type: string
          nullable: true
          maxLength: 100
        backstory:
          type: string
          nullable: true
          maxLength: 5000
        system:
          type: string
          nullable: true
          maxLength: 4000
          description: Canonical Eliza system prompt; systemPrompt remains as a compatibility alias.
        templates:
          type: object
          additionalProperties:
            type: string
          description: User-managed prompt template map. Backend templates are not exposed.
        settings:
          $ref: '#/components/schemas/SafeCharacterSettings'
        bio:
          type: array
          items:
            type: string
            maxLength: 500
          minItems: 1
          maxItems: 10
        lore:
          type: array
          items:
            type: string
            maxLength: 500
          maxItems: 20
        topics:
          type: array
          items:
            type: string
            maxLength: 50
          maxItems: 30
        adjectives:
          type: array
          items:
            type: string
            maxLength: 30
          maxItems: 20
        style:
          $ref: '#/components/schemas/StyleConfig'
        messageExamples:
          type: array
          items:
            $ref: '#/components/schemas/ExampleMessage'
          maxItems: 20
        postExamples:
          type: array
          items:
            type: string
            maxLength: 280
          maxItems: 20
        systemPrompt:
          type: string
          nullable: true
          maxLength: 4000
          deprecated: true
          description: Compatibility alias for system.
        knowledge:
          type: array
          items:
            $ref: '#/components/schemas/KnowledgeDocument'
          maxItems: 5
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    UpdateAICharacterInput:
      type: object
      properties:
        name:
          type: string
          maxLength: 100
          description: WAGDIE-owned; exported but ignored on import.
        username:
          type: string
          nullable: true
          maxLength: 100
        backstory:
          type: string
          nullable: true
          maxLength: 5000
        system:
          type: string
          nullable: true
          maxLength: 4000
        templates:
          type: object
          nullable: true
          additionalProperties:
            type: string
        settings:
          $ref: '#/components/schemas/SafeCharacterSettings'
        bio:
          type: array
          items:
            type: string
            maxLength: 500
          minItems: 1
          maxItems: 10
        lore:
          type: array
          items:
            type: string
            maxLength: 500
          maxItems: 20
        topics:
          type: array
          items:
            type: string
            maxLength: 50
          maxItems: 30
        adjectives:
          type: array
          items:
            type: string
            maxLength: 30
          maxItems: 20
        style:
          $ref: '#/components/schemas/StyleConfig'
        messageExamples:
          type: array
          items:
            $ref: '#/components/schemas/ExampleMessage'
          maxItems: 20
        postExamples:
          type: array
          items:
            type: string
            maxLength: 280
          maxItems: 20
        systemPrompt:
          type: string
          nullable: true
          maxLength: 4000
          deprecated: true
          description: Compatibility alias for system.

    SafeCharacterSettings:
      type: object
      description: User-managed safe settings only; plugins, secrets, providers, runtime config, and migration metadata are backend-owned.
      properties:
        avatar:
          type: string
          nullable: true
          maxLength: 500
        metadata:
          type: object
          properties:
            wagdieUser:
              type: object
              nullable: true
              additionalProperties:
                nullable: true
                oneOf:
                  - type: string
                  - type: number
                  - type: boolean

    StyleConfig:
      type: object
      properties:
        all:
          type: array
          items:
            type: string
            maxLength: 200
          maxItems: 10
        chat:
          type: array
          items:
            type: string
            maxLength: 200
          maxItems: 10
        post:
          type: array
          items:
            type: string
            maxLength: 200
          maxItems: 10

    ExampleMessage:
      type: object
      required:
        - userMessage
        - assistantMessage
      properties:
        userMessage:
          type: string
          maxLength: 500
        assistantMessage:
          type: string
          maxLength: 1000

    KnowledgeDocument:
      type: object
      required:
        - id
        - filename
        - size
        - mimeType
        - uploadedAt
      properties:
        id:
          type: string
          format: uuid
        filename:
          type: string
        content:
          type: string
          description: Only included when explicitly requested
        size:
          type: integer
          maximum: 51200
        mimeType:
          type: string
          enum:
            - text/plain
            - text/markdown
        uploadedAt:
          type: string
          format: date-time

    ElizaCharacterExport:
      type: object
      required:
        - name
        - bio
        - lore
      properties:
        name:
          type: string
          description: WAGDIE-owned; import ignores this field with a warning.
        username:
          type: string
        backstory:
          type: string
          nullable: true
        system:
          type: string
          description: Canonical Eliza system prompt.
        templates:
          type: object
          additionalProperties:
            type: string
        settings:
          $ref: '#/components/schemas/SafeCharacterSettings'
        bio:
          type: array
          items:
            type: string
        lore:
          type: array
          items:
            type: string
        topics:
          type: array
          items:
            type: string
        adjectives:
          type: array
          items:
            type: string
        style:
          $ref: '#/components/schemas/StyleConfig'
        messageExamples:
          type: array
          items:
            type: array
            items:
              type: object
              properties:
                user:
                  type: string
                content:
                  type: object
                  properties:
                    text:
                      type: string
        postExamples:
          type: array
          items:
            type: string
        systemPrompt:
          type: string
          deprecated: true
          description: Compatibility alias for system; if both are imported and differ, system wins with a warning.
        knowledge:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              path:
                type: string
              content:
                type: string

    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object

    ValidationError:
      type: object
      required:
        - error
        - message
        - fields
      properties:
        error:
          type: string
          example: validation_error
        message:
          type: string
        fields:
          type: object
          additionalProperties:
            type: array
            items:
              type: string